// src/api/models/raceEnum.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Eternal Anvil (Backend)
 * The backend server for the Eternal Anvil.
 * OpenAPI spec version: v0.0.1
 */

/**
 * * `ELF` - Elf
* `DWARF` - Dwarf
 */
export type RaceEnum = typeof RaceEnum[keyof typeof RaceEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RaceEnum = {
  ELF: 'ELF',
  DWARF: 'DWARF',
} as const;