// src/api/models/economyEnum.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Eternal Anvil (Backend)
 * The backend server for the Eternal Anvil.
 * OpenAPI spec version: v0.0.1
 */

/**
 * * `MULTI` - Multi Resource
* `WOOD` - Single Wood
* `IRON` - Single Iron
* `CROP` - Single Crop
 */
export type EconomyEnum = typeof EconomyEnum[keyof typeof EconomyEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EconomyEnum = {
  MULTI: 'MULTI',
  WOOD: 'WOOD',
  IRON: 'IRON',
  CROP: 'CROP',
} as const;